Updated agent.py
Made all tensors be able to run on CUDA.
Changed the naming
Made a method for pushing to the memory, as to not clutter the main loop.
Updated the getAction func to add noise for exploration.
added a update_target_network function, which only updates the target networks parameters.
Made the update function more easy to read and follow when looking at the papers pseudocode.
Added methods to save and load models.